+++ src/api/user.ts
@@ -1,3 +1,5 @@
+import { Language } from "@spoke/spoke-codegen";
+
 /* eslint-disable import/prefer-default-export */
 export enum NotificationFrequencyType {
   All = "ALL",
@@ -5,3 +7,12 @@ export enum NotificationFrequencyType {
   Daily = "DAILY",
   None = "NONE"
 }
+
+export const languageEnumToLabel = (value: Language) => {
+  switch (value) {
+    case Language.Es:
+      return "Español";
+    default:
+      return "English";
+  }
+};