+++ src/containers/Login/index.tsx
@@ -1,10 +1,16 @@
+// 2025-05-25 this is NOT fully a functional component currently
+// https://github.com/With-the-Ranks/Spoke/issues/70
+import { Box } from "@material-ui/core";
+import { Language } from "@spoke/spoke-codegen";
 import { css, StyleSheet } from "aphrodite/no-important";
 import muiThemeable from "material-ui/styles/muiThemeable";
 import queryString from "query-string";
 import React from "react";
+import { withTranslation } from "react-i18next";
 import type { RouteChildrenProps } from "react-router-dom";
 import { withRouter } from "react-router-dom";
 import { compose } from "recompose";
+import LanguageSelector from "src/components/LanguageSelector";
 
 import theme from "../../styles/theme";
 import type { MuiThemeProviderProps } from "../../styles/types";
@@ -50,14 +56,12 @@ const styles = StyleSheet.create({
   }
 });
 
-const saveLabels: Record<string, string> = {
-  [UserEditMode.SignUp]: "Sign Up",
-  [UserEditMode.Login]: "Log In",
-  [UserEditMode.Reset]: "Save New Password",
-  [UserEditMode.RequestReset]: "Request Reset"
-};
-
-type LogalLoginProps = RouteChildrenProps & MuiThemeProviderProps;
+type LogalLoginProps = RouteChildrenProps &
+  MuiThemeProviderProps & {
+    // comes from withTranslation
+    t: (key: string) => string;
+    i18n: any;
+  };
 
 interface LogalLoginState {
   active: string;
@@ -77,8 +81,12 @@ class LocalLogin extends React.Component<LogalLoginProps, LogalLoginState> {
     };
   }
 
-  handleClick = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+  handleClick = (
+    e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
+    i18n: any
+  ) => {
     this.setState({ active: e.currentTarget.name });
+    i18n.changeLanguage(Language.En); // Reset language to English on login/signup
   };
 
   naiveVerifyInviteValid = (nextUrl: string) =>
@@ -86,7 +94,8 @@ class LocalLogin extends React.Component<LogalLoginProps, LogalLoginState> {
     nextUrl.includes("invite");
 
   render() {
-    const { location, history, muiTheme } = this.props;
+    const { location, history, muiTheme, i18n, t } = this.props;
+
     const rawNextUrl = queryString.parse(location.search).nextUrl;
     const nextUrl =
       (Array.isArray(rawNextUrl) ? rawNextUrl[0] : rawNextUrl) || "/";
@@ -100,6 +109,13 @@ class LocalLogin extends React.Component<LogalLoginProps, LogalLoginState> {
 
     const headerColor = muiTheme?.palette?.primary1Color ?? theme.colors.green;
 
+    const saveLabels: Record<string, string> = {
+      [UserEditMode.SignUp]: t("sign up"),
+      [UserEditMode.Login]: t("log in"),
+      [UserEditMode.Reset]: t("save new password"),
+      [UserEditMode.RequestReset]: t("request reset")
+    };
+
     return (
       <div className={css(styles.loginPage)}>
         {/* Only display sign up option if there is a nextUrl */}
@@ -109,30 +125,35 @@ class LocalLogin extends React.Component<LogalLoginProps, LogalLoginState> {
               className={css(styles.button)}
               type="button"
               name={UserEditMode.Login}
-              onClick={this.handleClick}
+              onClick={(e) => this.handleClick(e, i18n)}
               disabled={active === UserEditMode.Login}
             >
-              Log In
+              {t("log in")}
             </button>
             <button
               className={css(styles.button)}
               type="button"
               name={UserEditMode.SignUp}
-              onClick={this.handleClick}
+              onClick={(e) => this.handleClick(e, i18n)}
               disabled={active === UserEditMode.SignUp}
             >
-              Sign Up
+              {t("sign up")}
             </button>
           </section>
         )}
         <div className={css(styles.fieldContainer)}>
           <h2 className={css(styles.header)} style={{ color: headerColor }}>
             {active === UserEditMode.EmailReset
-              ? "Reset Your Password"
+              ? t("reset password")
               : active === UserEditMode.RequestReset
-              ? "Request a Password Reset Email"
-              : "Welcome to Spoke"}
+              ? t("request reset email")
+              : t("welcome")}
           </h2>
+          {active !== UserEditMode.SignUp && (
+            <Box mt={4}>
+              <LanguageSelector />
+            </Box>
+          )}
           {active === UserEditMode.Reset ? (
             <UserPasswordReset history={history} nextUrl={nextUrl} />
           ) : (
@@ -153,6 +174,7 @@ class LocalLogin extends React.Component<LogalLoginProps, LogalLoginState> {
 }
 
 const LocalLoginWrapper = compose<LogalLoginProps, unknown>(
+  withTranslation("Login"),
   muiThemeable(),
   withRouter
 )(LocalLogin);